extracted resource reader and added empty case data map test
