--- src/main/java/uk/gov/hmcts/reform/ucmc/model/Document.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.ucmc.model;
-
-public class Document {
-}
