+++ src/main/java/uk/gov/hmcts/reform/ucmc/utils/ResourceReader.java
@@ -1,0 +1,28 @@
+package uk.gov.hmcts.reform.ucmc.utils;
+
+import org.apache.commons.io.IOUtils;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+public class ResourceReader {
+
+    private ResourceReader() {
+        // NO-OP
+    }
+
+    public static String readString(String resourcePath) {
+        return new String(ResourceReader.readBytes(resourcePath));
+    }
+
+    public static byte[] readBytes(String resourcePath) {
+        try (InputStream inputStream = ResourceReader.class.getClassLoader().getResourceAsStream(resourcePath)) {
+            if (inputStream == null) {
+                throw new IllegalArgumentException("Resource does not exist");
+            }
+            return IOUtils.toByteArray(inputStream);
+        } catch (IOException ex) {
+            throw new IllegalArgumentException(ex);
+        }
+    }
+}
