+++ src/test/java/uk/gov/hmcts/reform/ucmc/model/CaseDataTest.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.ucmc.model;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.commons.io.IOUtils;
 import org.json.JSONException;
 import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
@@ -10,9 +10,9 @@ import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import java.io.IOException;
-import java.io.InputStream;
+import java.util.Map;
 
-import static uk.gov.hmcts.reform.ucmc.model.CaseDataTest.ResourceReader.readString;
+import static uk.gov.hmcts.reform.ucmc.utils.ResourceReader.readString;
 
 @SpringBootTest(classes = JacksonAutoConfiguration.class)
 class CaseDataTest {
@@ -21,7 +21,7 @@ class CaseDataTest {
     private ObjectMapper mapper;
 
     @Test
-    void shouldCreateExpectedCaseData() throws IOException, JSONException {
+    void shouldCreateExpectedCaseDataWhenReadingFromJson() throws IOException, JSONException {
         String data = readString("case_data.json");
         CaseData caseData = mapper.readValue(data, CaseData.class);
         String dataAfterMapping = mapper.writeValueAsString(caseData);
@@ -29,26 +29,11 @@ class CaseDataTest {
         JSONAssert.assertEquals(data, dataAfterMapping, true);
     }
 
-    //TODO: move to own class
-    public static class ResourceReader {
+    @Test
+    void shouldNotErrorWhenMappingEmptyCaseData() throws JsonProcessingException, JSONException {
+        CaseData caseData = mapper.convertValue(Map.of(), CaseData.class);
+        String dataAfterMapping = mapper.writeValueAsString(caseData);
 
-        private ResourceReader() {
-            // NO-OP
-        }
-
-        public static String readString(String resourcePath) {
-            return new String(ResourceReader.readBytes(resourcePath));
-        }
-
-        public static byte[] readBytes(String resourcePath) {
-            try (InputStream inputStream = ResourceReader.class.getClassLoader().getResourceAsStream(resourcePath)) {
-                if (inputStream == null) {
-                    throw new IllegalArgumentException("Resource does not exist");
-                }
-                return IOUtils.toByteArray(inputStream);
-            } catch (IOException ex) {
-                throw new IllegalArgumentException(ex);
-            }
-        }
+        JSONAssert.assertEquals(Map.of().toString(), dataAfterMapping, true);
     }
 }
