+++ src/test/resources/sample-resource.txt
@@ -1,0 +1,1 @@
+Sample content
